backend-python/requirements.txt
flask
PyMySQL
flask_restful
flask-sqlalchemy
flask_migrate
flask_marshmallow
flask-cors
marshmallow-sqlalchemy
pytest
coverage